Adiciona tipo de usuário Admin
