+++ src/controllers/AdminController.js
@@ -1,0 +1,31 @@
+const con = require('../database');
+
+module.exports = {
+    index: async (req, res, next) => {
+        try {
+            const { id, page = 1 } = req.query
+            const query = await con('usuario').where({'desativado_em': null})
+            .limit(10)
+            .offset((page - 1) * 10)
+            .orderBy('id')
+
+            const countObj = con('usuario').count()
+
+            if (id) {
+                query
+                .where({ id })
+                .select('usuario.email')
+            }
+
+            const [count] = await countObj
+            res.header('X-Total-Count', count["count"])
+
+            const results = await query
+
+            return res.json(results);
+
+        } catch (error) {
+            next(error)
+        }
+    }
+}
