+++ src/controllers/UsuarioController.js
@@ -29,33 +29,6 @@ module.exports = {
             next(error)
         }
     },
-    index: async (req, res, next) => {
-        try {
-            const { id, page = 1 } = req.query
-            const query = await con('usuario').where({'desativado_em': null})
-            .limit(10)
-            .offset((page - 1) * 10)
-            .orderBy('id')
-
-            const countObj = con('usuario').count()
-
-            if (id) {
-                query
-                .where({ id })
-                .select('usuario.email')
-            }
-
-            const [count] = await countObj
-            res.header('X-Total-Count', count["count"])
-
-            const results = await query
-
-            return res.json(results);
-
-        } catch (error) {
-            next(error)
-        }
-    },
     read: async (req, res, next) => {
         try {
             const authHeader = req.headers.authorization
